Add a super disgusting way of passing values to interpreted haskell
In addition to being utterly disgusting I'm pretty sure it's not safe to
do it like this. Maybe performing the IO at assignment would help but
I'm not actually sure how to do that (need to see what ghci does).
Another option to generate an IO ref with the interpreter and have a
function from the ffi that lets me fill an IO ref with the value of a
stable pointer. It's a bit of faffing but I think it's a 'more correct'
way of doing things.